pull out pointers
